AudioZone.cs + AudioHandler.cs updates
AudioZoneSphere.cs changed to > AudioZone.cs
AudioZone has a separate editor script to prevent errors when building the game.
AudioZone.cs does not produce errors  anymore when: The AudioHandler isn't present, When the track doesn't exist and when it can't find an AudioListener in the scene.

AudioHandler.cs: Added extra checks to prevent errors

Added AudioZoneEditor